error-profile: Implementing linear regression

+++ error-profile/src/main/java/com/hartwig/hmftools/errorprofile/ReadBaseSupport.java
@@ -23,7 +23,7 @@ public class ReadBaseSupport
         public int readPosition5To3;
         public int refPosition;
 
-        public int baseQuality;
+        public byte baseQuality;
 
         public byte[] trinucleotideContext;
